# repo/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hospitality Operations Management System API
  description: Backend API for managing events, resources, reservations, recipes, inventory, entitlements, and operational workflows.
  version: 1.0.0
servers:
  - url: /api
    description: API base path
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from POST /auth/login or POST /auth/refresh.
  schemas:
...
    Resource:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        category_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [draft, published, unpublished]
        attributes:
          type: object
        created_at:
          type: string
          format: date-time
...
paths:
...
  /catalog/resources:
    get:
      tags: [Catalog]
      summary: List resources
      parameters:
        - name: category_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, unpublished]
      responses:
        '200':
          description: Resources
    post:
      tags: [Catalog]
      summary: Create a resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, category_id]
              properties:
                name:
                  type: string
                category_id:
                  type: string
                  format: uuid
                attributes:
                  type: object
      responses:
        '201':
          description: Resource created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
  /catalog/resources/{id}:
    get:
      tags: [Catalog]
      summary: Get resource details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Resource details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
    put:
      tags: [Catalog]
      summary: Update resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                attributes:
                  type: object
      responses:
        '200':
          description: Resource updated
  /catalog/resources/{id}/publish:
    post:
      tags: [Catalog]
      summary: Publish a resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Resource published
...
  /catalog/templates:
    get:
      tags: [Catalog]
      summary: List resource templates
      responses:
        '200':
          description: Templates
    post:
      tags: [Catalog]
      summary: Create a resource template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, category_id, attributes_schema]
              properties:
                name:
                  type: string
                category_id:
                  type: string
                  format: uuid
                attributes_schema:
                  type: object
      responses:
        '201':
          description: Template created

  # ── Check-in ──────────────────────────────────────────
...